refactor(mcp): move MCP modules into src/main/mcp/ and tidy entry points
- Move all 8 mcp-*.ts files under src/main/mcp/ via git mv (history

  preserved); drop the mcp- prefix on the new filenames.

- Delete the placeholder mcp-test.ts (was a no-op left over from a

  sandbox subdir restriction).

- Rewrite intra-MCP imports to sibling paths and move diag-logger

  import to ../diag-logger; update main.ts and pty-manager.ts to

  the new ./mcp/<name> paths.

- Drop Ctrl+Shift+M keybinding (it clashed with the upstream tabMenu

  binding); the grants dialog is reachable via the tab right-click

  menu and the pane header overflow menu.
