--- src/main/mcp-test.ts
@@ -1,3 +1,0 @@
-// Unused placeholder. The MCP module lives in mcp-*.ts files at this level.
-// (The session sandbox doesn't permit creating subdirectories.)
-export {};
