+++ src/main/mcp/copilot-autoreg.ts
